improved for deleting post

+++ Rentit-frontend/src/pages/userDetails/productCard.jsx
@@ -1,4 +1,5 @@
 import { useNavigate } from 'react-router-dom';
+import { useState } from 'react';
 import { CgRemove } from 'react-icons/cg';
 import { RiDeleteBinLine } from 'react-icons/ri';
 import axios from 'axios';
@@ -6,7 +7,9 @@ import axios from 'axios';
 
 const ProductCard = (props) => {
   const navigate = useNavigate();
-  const data = props.data;
+  const [data, setData] = useState(props.data);
+  const [count, setCount] = useState(props.count);
+  // const data = props.data;
   const type = props.type;
   const user = JSON.parse(sessionStorage.getItem('userDetails')).email;
   const url = 'http://127.0.0.1:2000/api-rentit/v1/items';
@@ -14,8 +17,17 @@ const ProductCard = (props) => {
     navigate('/itemDetails', { state: { id: id } });
   };
   const removeProduct = async (id, type) => {
-    await axios.delete(`${url}/${id}/${user}/${type}`);
+    if (window.confirm('Do you want do delete this post?')) {
+      await axios.delete(`${url}/${id}/${user}/${type}`);
+      setData(null);
+      setCount(count - 1);
+    }
   };
+  if (!data) {
+    if (!count && type === 'posts') return <>No produts posted</>;
+    else if (!count && type === 'saves') return <>No products saved</>;
+    else return <></>;
+  }
   return (
     <>
       <div className='cellBox'>
